Move and rename.
